{"schema":"decodex/commit/1","summary":"Extract frozen frame stream setup","authority":"manual"}
