+++ docs/reference/workspace-layout.md
@@ -192,8 +192,10 @@ The main host-kit files are split by responsibility:
   passthrough management
 - `CaptureOverlayImageSampler.swift`: CoreGraphics below-overlay capture and display-point sample
   adaptation for live chrome, frozen capture effects, and scroll fallback acquisition
-- `FrozenFrameAuthority.swift`: ScreenCaptureKit frozen-frame stream setup, self-capture-safe
-  stream gating, and lifecycle state
+- `FrozenFrameAuthority.swift`: frozen-frame authority state, frame storage, setup completion, and
+  telemetry bookkeeping
+- `FrozenFrameAuthority+StreamSetup.swift`: ScreenCaptureKit frozen-frame stream setup,
+  shareable-content lookup, self-capture-safe stream gating, and lifecycle reset
 - `FrozenFrameAuthority+SnapshotResolution.swift`: frozen-frame latch-token resolution, RGB/loupe
   sampling, self-capture-safe frame gating, and fresh-frame authority decisions
 - `FrozenFrameStreamOutput.swift`: ScreenCaptureKit stream-output delegate adaptation,
